+++ Tests/Linting/FileEncoding.Tests.ps1
@@ -10,19 +10,39 @@ Describe 'Linting\FileEncoding' {
                         'TestResults.xml'
                     );
 
-    Get-ChildItem -Path $repoRoot -Exclude $excludedPaths |
-        ForEach-Object {
-            Get-ChildItem -Path $_.FullName -Recurse |
-                ForEach-Object {
+    function TestEncodingPath {
+        [CmdletBinding()]
+        param (
+            [Parameter(Mandatory, ValueFromPipeline)]
+            [System.String] $Path,
 
-                    ## Resolve-ProgramFilesFolder.ps1 contains Unicode characters
-                    if (($_ -is [System.IO.FileInfo]) -and ($_.Name -ne 'Resolve-ProgramFilesFolder.ps1'))
+            [System.String[]] $Exclude
+        )
+        process
+        {
+            $WarningPreference = 'SilentlyContinue'
+            Get-ChildItem -Path $Path -Exclude $Exclude |
+                ForEach-Object {
+                    if ($_ -is [System.IO.FileInfo])
                     {
-                        It "File '$($_.FullName.Replace($repoRoot,''))' uses UTF-8 (no BOM) encoding" {
-                            $encoding = (Get-FileEncoding -Path $_.FullName -WarningAction SilentlyContinue).HeaderName
-                            $encoding | Should Be 'us-ascii'
+                        if ($_.Name -ne 'Resolve-ProgramFilesFolder.ps1')
+                        {
+                            It "File '$($_.FullName.Replace($repoRoot,''))' uses UTF-8 (no BOM) encoding" {
+                                $encoding = (Get-FileEncoding -Path $_.FullName -WarningAction SilentlyContinue).HeaderName
+                                $encoding | Should Be 'us-ascii'
+                            }
                         }
                     }
+                    elseif ($_ -is [System.IO.DirectoryInfo])
+                    {
+                        TestEncodingPath -Path $_.FullName -Exclude $Exclude
+                    }
                 }
+        } #end process
+    } #end function
+
+    Get-ChildItem -Path $repoRoot -Exclude $excludedPaths |
+        ForEach-Object {
+            TestEncodingPath -Path $_.FullName -Exclude $excludedPaths
         }
 }