+++ ChangeLog.md
@@ -8,7 +8,7 @@
 * Updates bundled xPendingReboot DSC resource to v0.4.0.0
 * Fixes a bug where a certain BitLocker policy causes mounting errors
 * Switches from PowerShell.org community build server to Appveyor
-* Adds PSScriptAnalyzer linting tests
+* Adds PSScriptAnalyzer and styling linting tests
 
 ### v0.17.0 ###